requirements.txt
# Create requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
langchain>=0.1.0
langchain-openai>=0.1.0
langserve>=0.1.0
langchain-community>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0  # SQLite ORM
bcrypt>=4.0.0  # Auth
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib>=1.7.4  # Password hashing